Move collate_fn to trainer_utils
